+++ sdk/storage/azure-storage-files-datalake/src/datalake_file_system_client.cpp
@@ -203,7 +203,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     }
     else if (options.AccessType == Models::PublicAccessType::None)
     {
-      blobOptions.AccessType = Blobs::Models::PublicAccessType::Private;
+      blobOptions.AccessType = Blobs::Models::PublicAccessType::None;
     }
     else
     {
@@ -340,7 +340,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     {
       ret.AccessType = Models::PublicAccessType::Path;
     }
-    else if (result->AccessType == Blobs::Models::PublicAccessType::Private)
+    else if (result->AccessType == Blobs::Models::PublicAccessType::None)
     {
       ret.AccessType = Models::PublicAccessType::None;
     }
@@ -376,7 +376,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     }
     else if (options.AccessType == Models::PublicAccessType::None)
     {
-      blobOptions.AccessType = Blobs::Models::PublicAccessType::Private;
+      blobOptions.AccessType = Blobs::Models::PublicAccessType::None;
     }
     else
     {