+++ sdk/storage/azure-storage-blobs/CHANGELOG.md
@@ -13,6 +13,7 @@
 - `ListType` in `GetBlockListOptions` was changed to non-nullable.
 - Added `BlobLeaseClient`, all lease related APIs are moved to `BlobLeaseClient`.
 - Type for lease duration in requests was changed to `std::chrono::seconds`, in response was changed to enum.
+- `PublicAccessType::Private` was renamed to `PublicAccessType::None`.
 - `startsOn` parameter for `GetUserDelegationKey` was changed to optional.
 
 ## 12.0.0-beta.6 (2020-01-14)